requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
APScheduler==3.10.1
asgiref==3.7.2
async-timeout==4.0.2
async-tio==1.3.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
curl-cffi==0.5.6
Django==3.2.2
django-cors-headers==4.0.0
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
dnspython==2.3.0
exceptiongroup==1.1.1
Flask==2.3.2
frozenlist==1.3.3
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mongoengine==0.27.0
multidict==6.0.4
openai==0.27.7
OpenAIAuth==1.0.2
Pillow==9.5.0
prompt-toolkit==3.0.38
pycparser==2.21
PyJWT==1.7.1
pymongo==4.3.3
PySocks==1.7.1
pytz==2023.3
regex==2023.5.5
requests==2.31.0
revChatGPT==6.0.1
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
sqlparse==0.4.4
tiktoken==0.4.0
tqdm==4.65.0
typing_extensions==4.6.2
tzdata==2023.3SS
tzlocal==5.0.1
urllib3==2.0.2
wcwidth==0.2.6
Werkzeug==2.3.4
yarl==1.9.2
